разделил код Engine.ts по файлам
