+++ src/engine/MapObject.ts
@@ -1,0 +1,7 @@
+import { Sprite } from "./Sprite";
+
+export class MapObject extends Sprite {
+  collide(object: MapObject) {
+    this.emit("collision", object);
+  }
+}
